More work on getting the jhbuild more streamlined
git-svn-id: http://developer.imendio.com/svn/gtk-osx-build@691 62e0af8c-05e7-0310-8efa-89bbc38b0f4f

+++ gtk-osx-build-setup.sh
@@ -13,7 +13,7 @@ else
 fi
 
 echo "Installing jhbuild..."
-(cd jhbuild && make -f Makefile.plain install >/dev/null)
+(cd jhbuild && make -f Makefile.plain DISABLE_GETTEXT=1 install >/dev/null)
 
 echo "Checking out gtk-osx build setup from subversion..."
 if ! test -d $SOURCE/gtk-osx-build; then
@@ -25,7 +25,9 @@ fi
 echo "Installing gtk-osx build setup..."
 (cd gtk-osx-build; ln -sfh `pwd`/jhbuildrc-gtk-osx $HOME/.jhbuildrc)
 (cd gtk-osx-build; ln -sfh `pwd`/jhbuildrc-gtk-osx-fw-10.4 $HOME/.jhbuildrc-fw-10.4)
-(cd gtk-osx-build; cp jhbuildrc-gtk-osx-custom-example $HOME/.jhbuildrc-custom)
+if [ ! -f $HOME/.jhbuildrc-custom ]; then
+    (cd gtk-osx-build; cp jhbuildrc-gtk-osx-custom-example $HOME/.jhbuildrc-custom)
+fi
 
 if test "x$1" == "xmaintainer-mode"; then
     echo "Setting up extra modulesets for maintainer mode..."
